# openapi.yaml
openapi: "3.1.0"
info:
  title: Malluland API
  version: "1.0.0"
  description: Official API for Malluland backend services.
servers:
  - url: http://localhost:8787
    description: Local server
components:
  securitySchemes:
    XUserId:
      type: apiKey
      in: header
      name: x-user-id
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
paths:
...
  /admin/logs/create:
    post:
      summary: Create admin log entry
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [userId, actionType]
              properties:
                userId: { type: string, format: uuid }
                actionType: { type: string }
                cmsPage: { type: string, nullable: true }
      responses:
        '200': { description: Created }
  /admin/logs/list:
    get:
      summary: List admin logs
      security:
        - bearerAuth: []
      parameters:
        - name: page
          in: query
          schema: { type: integer, minimum: 1, default: 1 }
        - name: limit
          in: query
          schema: { type: integer, minimum: 1, default: 50 }
      responses:
        '200':
          description: List
          content:
            application/json:
              schema:
                type: object
                properties:
                  data: { type: array, items: { type: object } }
                  metadata:
                    type: object
                    properties:
                      page: { type: integer }
                      size: { type: integer }
                      count: { type: integer }
  /admin/logs/{id}:
    get:
      summary: Get admin log by id
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      responses:
        '200': { description: Ok }
        '404': { description: Not found }
...
  /admin/avatar-updates:
    get:
      summary: List pending avatar updates
      security:
        - bearerAuth: []
      parameters:
        - name: page
          in: query
          schema: { type: integer, minimum: 1, default: 1 }
        - name: pageSize
          in: query
          schema: { type: integer, minimum: 1, default: 10 }
      responses:
        '200': { description: Ok }
  /admin/avatar-updates/approve:
    post:
      summary: Approve avatar photos
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [ids]
              properties:
                ids: { type: array, items: { type: string, format: uuid } }
      responses:
        '200': { description: Ok }
  /admin/avatar-updates/reject:
    post:
      summary: Reject avatar photos
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [ids]
              properties:
                ids: { type: array, items: { type: string, format: uuid } }
      responses:
        '200': { description: Ok }
...
  /chats/rooms:
    get:
      summary: List chat rooms for the authenticated user
      description: Returns the chat rooms in which the caller participates, including unread count and the last message metadata.
      security:
        - XUserId: []
      parameters:
        - in: query
          name: page
          required: false
          schema:
            type: integer
            default: 1
            minimum: 1
      responses:
        '200':
          description: A paginated list of chat rooms
          content:
            application/json:
              schema:
                type: object
                properties:
                  page:
                    type: integer
                  pageSize:
                    type: integer
                  total:
                    type: integer
                  items:
                    type: array
                    items:
                      type: object
                      properties:
                        id:
                          type: string
                          format: uuid
                        type:
                          type: string
                          enum: [DM, meetup]
                        meetup_id:
                          type: string
                          format: uuid
                          nullable: true
                        unread_count:
                          type: integer
                        last_message_id:
                          type: string
                          format: uuid
                          nullable: true
                        last_message_kind:
                          type: string
                          nullable: true
                          description: text | system
                        last_message_body:
                          type: string
                          nullable: true
                        last_message_at:
                          type: string
                          format: date-time
                          nullable: true
                        participant_user_ids:
                          type: array
                          items:
                            type: string
                            format: uuid
        '401':
          description: Unauthorized
...
  /chats/rooms/{id}/send:
    post:
      summary: Send a message to a chat room
      security:
        - XUserId: []
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string, format: uuid }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [text]
              properties:
                text: { type: string, minLength: 1 }
      responses:
        '201':
          description: Created chat message with participants
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: object
                    properties:
                      id: { type: string, format: uuid }
                      chatId: { type: string, format: uuid }
                      senderUserId: { type: string, format: uuid }
                      kind: { type: string, enum: [text, system] }
                      body: { type: string }
                      createdAt: { type: string, format: date-time }
                  participants:
                    type: array
                    items:
                      type: object
                      properties:
                        userId: { type: string, format: uuid }
                        name: { type: string, nullable: true }
                        avatarUrl: { type: string, nullable: true }
        '400': { description: Bad request }
        '401': { description: Unauthorized }
...